feat : vote 추가

+++ server/stackoverflow/src/main/java/com/idle/stackoverflow/answer/dto/AnswerDto.java
@@ -38,7 +38,7 @@ public class AnswerDto {
         private String content;
         private LocalDateTime createdAt;
         private LocalDateTime modifiedAt;
-        private long voteCnt;
+        private int voteCnt;
         private long userId;
         private long questionId;
     }
